The DefaultComparerMin for float should use Comparer<float>

+++ Funcky/Extensions/EnumerableExtensions/MinOrNone.cs
@@ -328,7 +328,7 @@ namespace Funcky.Extensions
 
         // We impose a total order where NaN is smaller than NegativeInfinity (same behaviour as dotnet)
         private static float DefaultComparerMin(float left, float right)
-            => Comparer<double>.Default.Compare(left, right) < 0 ? left : right;
+            => Comparer<float>.Default.Compare(left, right) < 0 ? left : right;
 
         // We impose a total order where NaN is smaller than NegativeInfinity (same behaviour as dotnet)
         private static double DefaultComparerMin(double left, double right)
